requirements.txt
asgiref==3.7.2
certifi==2023.7.22
charset-normalizer==3.2.0
Django==4.2.4
django-rest==0.8.7
djangorestframework==3.14.0
djangorestframework-jwt==1.11.0
idna==3.4
PyJWT==1.7.1
pytz==2023.3
requests==2.31.0
six==1.16.0
sqlparse==0.4.4
typing_extensions==4.7.1
tzdata==2023.3
urllib3==2.0.4
